+++ src/components/Auth/validateMessages.ts
@@ -1,0 +1,8 @@
+const validateMessages = {
+  required: 'This field is required!',
+  types: {
+    email: 'This is not a valid email!',
+  },
+};
+
+export default validateMessages;
